chore: enable address sanitizer in debug builds
Although C++ is a powerful language, it's flexibility opens the door to
several classes of memory issues which can result in undefined behavior
and program instability.

Address sanitizer (ASan) is an essential tool for finding such memory
vulnerabilities. Unfortunately, it does incur additional runtime cost
which is why it's only enabled for debug builds.

Closes #2