feat: create a window using GLFW
The first step for any renderer is to display a window. Unfortunately,
each operating system does this differently and there is no standard
solution for achieving this.

To facilitate cross-platform window creation, a library known as GLFW is
employed which abstracts the implementations details needed to display a
window. It also provides APIs for input handling for both keyboard/mouse
and gamepad controllers. Although it was primarily developed for OpenGL,
it does have explicit support for Vulkan.

Another library that was considered was SDL2 which has the added
benefits of supporting more platforms such as iOS and Android in
addition to audio support. However, the library contains many other
unnecessary features and is poorly documented compared to GLFW. In an
effort to be lean, this project has opted to go with GLFW and can
reconsider SDL when version 3 is released.

Window creation was confirmed to work on both Windows and Ubuntu using
Windows Subsystem for Linux.

More information about GLFW can be found here: https://www.glfw.org.