docs: add pull request policy
To ensure a high standard of code quality, this project has enabled
branch protection policies for main on GitHub.

To ensure contributors are aware of and follow this policy a relevant
section has been added to the README. This was also used as an
opportunity to revise existing sections to me more concise, clear, and
add additional information about the vcpkg.json manifest.