feat: create a Vulkan instance
There is no global state in Vulkan and all per-application state is
stored in a VkInstance object. Creating a VkInstance object initializes
the Vulkan library and allows the application to pass information about
itself to the implementation.

Although trivial to create, initializing a Vulkan instance requires
environment configuration to utilize Vulkan libraries. To achieve this,
Vulkan-Hpp was added which, in addition to C++ bindings for the Vulkan C
API, provides an implementation for a dynamic dispatch loader to load
Vulkan APIs at runtime.

This commit also enables the LunarG validation layers for debug builds.
As a result, the Vulkan SDK is currently required for debug builds
(although it is not necessary for release builds because Vulkan-Hpp is
managed by vcpkg). This requires a README update which is not included
in this commit and is tracked by #21.

Relevant documentation:
- https://www.glfw.org/docs/3.3/vulkan_guide.html
- https://github.com/KhronosGroup/Vulkan-Hpp#extensions--per-device-function-pointers
- https://registry.khronos.org/vulkan/specs/1.3-extensions/man/html/vkCreateInstance.html

Closes #20.