style: update clang-format rules
There are sections in the Google C++ Style Guide which do not mandate a
particular style other than maintaining consistency with existing
conventions in the file. For example, both int* x and int *x are allowed
in addition to const int x and int const x.

To ensure a consistent style throughout the project, clang-format
options are being updated to enforce a single style for both of the
previously described scenarios. Additionally, the README has been
updated to include exceptions to the Google C++ Style Guide that are
used in this project.

Inspiration for these changes come from:
https://chromium.googlesource.com/chromium/src/+/main/styleguide/c++/c++.md
https://github.com/googleapis/google-cloud-cpp/blob/main/.clang-format