fix: don't invoke empty std::function
If a window's on key event handler is not set, an exception will be
thrown due to trying to invoke an empty std::function.

The trivial fix for this is to invoke std::function's implicit boolean
conversion operator. This change also adds a change that uses
std::assert to verify the GLFW user pointer rather than a runtime check
since the logic is internal to the window class.