feat: get Vulkan device queues
Vulkan queues represent the primary abstraction for sending commands to
the GPU.

During the physical device selection process, queue indices are
identified and cached which are then used during the device creation
process to specify which queue families should be used. After the
logical device has been created, getting a Vulkan queue is a simple
process involving one API call with the appropriate queue family and
queue indices.

Despite the trivial nature of getting a Vulkan queue, the existing
design made it difficult to achieve due to the requirement to cache
queue family indices during the physical device selection process. This
precludes initializing a gfx::Queue in the member initialization list
without awkward use of forwarding arguments as a std::tuple. Although
it's possible to achieve this in the constructor body, a default
constructor for gfx::Queue was specifically avoided to enforce the
class invariant of having member variables be fully initialized when
later accessed with class APIs.

To mitigate this, a private constructor was added that uses the internal
RankedPhysicalDevice which allows initializing everything in the member
initialization list. This unfortunately requires exposing
RankedPhysicalDevice as a nested subclass however this was mitigated by
both forward declaring the type and making it private to gfx::Device.
Note that this can be fixed when the project is transitioned to
C++20 modules which has first class support for controlling which parts
of the API are exported.

Relates to #25