chore: fix x86 builds
x86 debug builds are failing with an error indicating that a
"configuration files were considered but not accepted". Upon further
investigation, it was determined that vcpkg was downloading x86 packages
but attempting to compile them with x64 version of cl.exe. This was due
to a bug in CMakePresets.json which did not correctly merge
configuration values when inheriting multiple base configurations.

Originally, there were two base configuration presets, "ninja" and "x86"
which set "architecture.strategy" and "architecture.value" respectively.
Despite these being different, it appears CMake inherit rules only
considers conflicts for shallow properties resulting in the
"architecture" property being set to the first occurring "ninja"
configuration which left "architecture.value" to be undefined.
Interestingly, the behavior only seems to affect the "architecture"
property and not the "cacheVariables" property which was confirmed to
correctly merge CMake cache values when inherited from multiple base
configurations ¯\_(ツ)_/¯.

Due to this limitation in CMake, the fix for this requires splitting the
"ninja" configuration in two places: moving the "generator" property to
the root "default" configuration while moving the
"architecture.strategy" property to the "x64" and "x86" configurations.

To confirm the correct architecture, x64 and x86 buidls were made with
the following statements respectively:
- static_assert(sizeof(std::size_t) == sizeof(std::uint64_t))
- static_assert(sizeof(std::size_t) == sizeof(std::uint32_t))
which was confirmed to work in both debug and release builds for each
platform architecture.