fix: declare instance layers as constexpr
Previous attempts to make instance layers constexpr failed because of
limitations with zero-length arrays (in release mode) and constexpr
restrictions with std::vector.

This commit enables constexpr instance layers using an approach not
previously considered: std::initializer_list. This enables constexpr
lists without the restriction of hardcoding the number of elements which
can vary based on the target build type.