refactor: move implementation details out of header
Because gfx::Queue does not have a default constructor (to ensure class
invariants), gfx::Device requires a private constructor to forward
arguments which are calculated all at once so that they can be
initialized once in member initialization lists. This unfortunately
required exposing several private static methods that depended on the
internal RankedPhysicalDevice.

This commit reworks this design by removing the constraint of
RankedPhysicalDevice being private which enables moving functions
which depend on this type to an anonymous namespace in the cpp
implementation file. Although this is also suboptimal, it is considered
better than the previous design because it reduces the exposure class
implementation details in the gfx::Device header. Furthermore, this
change is more amenable to adoption of C++20 modules at which point
RankedPhysicalDevice can simply be excluded from the gfx::Device module
exports.