feat: enable model loading using assimp
Model loading is limited to a subset of the Wavefront .obj file format
specification which is currently restricted to loading vertex and index
data. This commit replaces the custom .obj loader implementation with
Asset Importer (a.k.a. Assimp) which support are large variety of modern
formats that enable future improvements related to animation and PBR.