refactor: move spherical coordinates to math lib
The logic for spherical coordinates currently lives inside of arc camera
which violates the single responsibility principle and requires the
unsavory practice of including a .cpp file to test internal logic.

This commit improves upon this by moving logic relating to spherical
coordinates to a dedicated file as part of a new math interface library.