feat: add texture support
Texture mapping is an essential technique in computer graphics that
enables applying an image to an object surface using coordinates inside
an image region.

Achieving this in Vulkan is somewhat involved as it requires both
allocating enough descriptor sets for each unique material in addition
to configuring the the graphics pipeline to declare texture bindings
upfront which must align with bindings specified in shaders and the
types of materials available in an model. Ideally, this would be
determined dynamically by analyzing what materials are available in a
model and generating appropriate shaders + descriptor layouts to render
it. However, this is a significant amount of work that is outside the
scope of this commit so this incurs some technical debt by hardcoding
expectations about what materials are available in a model when
rendering.