style: remove cpplint
Although cpplint somewhat helps with ensuring a consistent code style,
it is largely superseded by more modern tools such as clang-format and
clang-tidy. Furthermore, it is not regularly updated and therefore has
many false positives with modern C++ features that requires littering
the codebase with NOLINT comments which adds noise and degrades
readability.