refactor: simplify device creation
One of the primary limitations with the current design for physical
device selection and subsequent logical device creation is leaking
internal implementation details (i.e, RankedPhysicalDevice) which is
ultimately due to not having a default constructor for the queue family
abstraction which was explicitly omitted to preserve class invariants.

This commit reworks the design to avoid leaking this internal construct
by including constructing a physical device abstraction during the
ranking process rather than afterwards. This commit also removes the
abstractions for queue and queue family in favor of a simple wrapper
around queue family indices which greatly simplifies the design.