build: add fast math compiler options
Fast math enables a set of compiler optimizations that are allowed to
break strict IEEE floating point standards such as reording, combining,
or simplifying operations to improve both speed and space. This is often
a required for enabling SIMD operations which is critical for speeding
up CPU based vector/matrix operations. It's worth noting this comes at
the cost of potentially unsafe optimizations that can introduce rounding
errors for example, but this is generally acceptable outside fields that
require strict floating point precision.

Closes #57