refactor: remove data view abstraction
The purpose of the data view abstraction was to create a span like
structure that can also represent a single value. Vulkan-Hpp also has
an abstraction for this called vk::ArrayPoxy which can be used instead
to avoid the maintenance liability. This commit also replaces excessive
uses of data view with std::span when it is guaranteed that an argument
will only ever been an array like structure.