feat: render the sponza glTF model
The goal of this commit is to expand the project scope from a simple
model viewer to one that can explore more complex scenes. To achieve
this, the arc camera implementation is replaced with a free camera and
the model is replaced with the Sponza Castle glTF model from the Khronos
glTF samples repository. Unfortunately, this model does not include
normal maps for all of its meshes and because this project not yet
support dynamic descriptor set layout generation, normal mapping had to
be temporarily disabled.