build: generate JSON compilation database
This commit enables the CMAKE_EXPORT_COMPILE_COMMANDS variable for ninja
generated builds to create a JSON compilation database. This is useful
for static analyzers found in Clang Libtools that depend on the abstract
syntax tree but may not be part of the core build system.

See the following document for more information:
https://clang.llvm.org/docs/JSONCompilationDatabase.html