build: reinstate cpplint
Although imperfect, CppLint is a valuable static anlaysis tool for
enforcing a consistent style that adheres to the Google C++ Style Guide.
It was previously removed because it is largely superseded by clang-tidy
however there isn't a reliable way to integrate this on Windows (see #4
for more details). Therefore, CppLint is being reinstated until a better
alternative becomes available.