feat: submit copy commands to the transfer queue
Some physical devices have hardware support for a dedicated transfer
queue which enables asynchronous asset uploading using Direct Memory
Access (DMA).

This commit implements the first step in realizing asynchronous asset
uploading by searching for a dedicated transfer queue during the
physical device selection process and using that queue to upload buffer
or image data. Note that this process still waits for the queue to be
idle before returning control back to the caller so there is still work
to do in order to fully realize the benefits of this approach which is
tracked by #67.

Closes #49