feat: manage memory using Vulkan Memory Allocator
One of the strengths of Vulkan is that it gives the application full
control over memory management. However, doing this effectively requires
implementing a custom memory allocator which adds significant complexity
to the project. Fortunately, AMD has a created Vulkan Memory Allocator
(VMA) which is a battle tested memory allocator used commercially in
many games and professional graphics software.

This commit replaces the current rudimentary memory allocation strategy
with VMA to manage memory of buffers and images. Furthermore,
incorporating VMA into the project has led to a 26% performance
improvement in load times when importing the Sponza glTF model.

Closes #58