style: avoid nested classes
This project makes use of nested classes to avoid "polluting" the gfx
namespace. This is generally considered bad practice because namespaces
are better suited for this purpose. At this time, it's not necessary to
introduce new namespaces given no naming conflicts. Additionally, adding
nested namespaces has resulted in compiler errors when experimenting
with transitioning to C++20 modules.