refactor: enable async asset loading
Asset loading is currently a performance bottleneck when starting the
application which is partly due to synchronously uploading assets to the
GPU. To address this, buffers are loaded upfront with a single command
buffer recording each copy command which is then submitted to a transfer
queue while the remainder of the scene finishes importing.

This commit also includes a minor improvement to enable configuring the
memory priority for dedicated image attachments to prevent the OS from
evicting them from memory. Collectively, these performance improvements
have resulted in as much as a 12% improvement in start times. Lastly,
this also will enable future performance enhancements by applying the
same pattern to loading materials.

Closes #67