refactor: add camera translation using WASD keys
The current mechanism for translating the camera uses the right mouse
button and scroll wheel. While appropriate for a simple model viewer
this is less ideal for exploring environments such as Sponza.

This commit refactors the current input mechanism to use WASD keys to
control camera translation. To achieve this, the logic for handling
input was refactored from a hybrid event-based/polling solution to
exclusively poll for input events given to maintain simplicity. Although
event-based input is more elegant, it added significant complexity that
could not be justified when only 4 keys and one mouse button are used.