style: remove redundant brace initialization
This is a minor commit which removes redundant brace initialization when
the type in question is known to have a default constructor. This was
previously added to ensure member variables are always initialized, but
it introduced inconsistency with other forms of variable initialization.
The decision moving forward is to only require value-initialization for
primitive types, enums, or structs without a user defined constructor.