refactor: replace assimp with cgltf
Although assimp has a user-friendly API that supports multiple file
formats, it's runtime performance is significantly slower than other
alternatives. Furthermore, its compile times are truly abysmal with a
clean build using vcpkg taking several minutes.

Because this project exclusively used glTF files, it is more appropriate
to use a lightweight library for handling those such as cgltf. Although
cgltf is more complicated (primarily due to needing to understand the
glTF specification), it is also significantly faster. When compared to
assimp (with no preprocessing flags), load times were reduced by 89.8%!
Additionally,  load times are also much, much faster due to cgltf being
header only which should be a huge productivity booster.

The only downside to this approach is that it restricts this project to
glTF files. With the exception of Universal Scene Description (which
isn't even supported by assimp), it's not anticipated that this project
will support any other file formats so the tradeoff is justified given
the benefits previously described.