style: reorder parameters
Parameter ordering has generally followed the same strategy of passing
the Vulkan device as the first parameter. However, often times this
isn't the most important parameter and is instead simply an auxiliary
type used to create the object in question. This commit reorders
parameters so that types directly relevant to the object creation appear
first.