refactor: add load model API
Martin Fowler coined the term "Tell Don't Ask" to describe a paradigm of
telling an object invoke some behavior rather than asking for the data
to perform the same action by the client. This is a useful principle
because it encourages encapsulating data and logic that would other leak
outside that class.

One issue with the engine class is that it exposes the Vulkan memory
allocator for use with creating models rather than creating an API to
tell the engine to load a model with said allocator. By refactoring this
exposing the allocator through a getter can be avoided in addition to a
few other code cleanup improvements.