refactor: avoid leaking implementation details
Despite gfx::Node being used internally in gfx::Model, it is defined
outside that class because it needs to be publicly visible to internal
methods defined in the implementation file. This is suboptimal because
it leaks implementation details and needlessly pollutes the surrounding
namespace.

One solution to fix this is to define gfx::Node inside the private
access scope of gfx::Model and convert internal methods that use it from
the anonymous namespace to static member variables. Although this
addresses polluting the surrounding namespace, it still ends up leaking
implementation details (e.g., exposing cgltf types). A better approach
is to convert those internal methods to methods of gfx::Model::Node
which enables forward declaring that class so that implementation
details can remain hidden in the implementation file. One caveat to this
approach is that it requires defining the previously compiler generated
gfx::Model destructor in the implementation file so that std::unique_ptr
has access to the gfx::Model::Node destructor.