refactor: remove uniform buffer objects
While the current approach of using uniform buffers to pass camera
transformations is valid, it adds unnecessary complexity when they can
instead be passed using push constants which is guaranteed to contain
at least 128 bytes by the Vulkan specification.

The primary motivation for this is to simplify the implementation in
preparation for a larger refactor to have the graphics pipeline
generation driven by the model rather than engine which in turn will
simplify the process for creating descriptor sets for model textures.