refactor: verify required vertex attributes
When processing mesh primitives, required vertex attributes are not
verified to exist prior to constructing a mesh vertex. To address this,
unpacked vertex attributes are wrapped in a std::optional whose values
are later accessed with std::optional::value which will throw a
std::bad_optional_access exception if the vertex attribute was never set
due to it existing in the glTF file.