refactor: rework parameter ordering
This commit reworks parameter ordering to be more consistent with the
core Vulkan API (e.g., having vk::Device appear as the first parameter).

This commit also simplifies gfx::Window to use Vulkan and glm types in
its API signatures which avoids awkward type conversion in the name of
theoretically not using Vulkan in the future.