refactor: add physical device abstraction
As this project evolves the need to repeatedly access physical device
properties is becoming more frequent. Because repeatedly quering the
driver for this information is inefficient, this data needs to be stored
somewhere. While this could be added to gfx::Device along with
vk::PhysicalDevice, it makes better sense to create a physical device
abstraction which can cache physical device properties and features.