feat: add anisotropic filtering
In computer graphics, undersampling arises when there are more texels
than fragments which can result in artifacts when viewing textures from
oblique angles.

One technique to help address this is ansiotropic filtering which can be
applied automatically by a sampler in Vulkan. To achieve this, the
sampler anisotropic feature must first be enabled when creating a
logical device and then later configured when creating a sampler.