build: remove unused textures
There are several textures that are not referenced in the main Sponza
glTF file that can be safely removed. These mostly include individual
metal and roughness textures which are instead referenced in a combined
metal/roughness texture. There were some references to curtains which
appear to have been moved to a separate model.