fix: validate depth attachment format support
The depth buffer attachment is always created with the eD24UnormS8Uint
format regardless of whether or not it's supported on the target
platform which can result in an application crash. To fix this,
candidate depth attachment formats are evaluated using physical device
format properties to ensure that a format is supported prior to creating
a depth attachment.