style: update std::optional naming convention
For variables that represent a std::optional<T>, it's often difficult to
disambiguate the the underlying type by name alone. One convention that
is starting to emerge in the C++ community is to prefix std::optional
types with "maybe_" which better indicates that the variable represents
an std::optional (whereas the a variable representing the underlying
value is the same name without maybe_).