refactor: enable additional clang-tidy rules
This commit removes the following clang-tidy exceptions:
  - readability-avoid-const-params-in-decls
  - readability-identifier-length
  - readability-isolate-declaration

readability-avoid-const-params-in-decls was initially disabled in commit
28fa970. However, upon further review this rule is reenabled because the
visual consistency of mixing const and non-const parameters in function
declarations is overstated and many guidelines recommend against it.

readability-identifier-length was initially disabled because variables
like x,y,z are well established in computer graphics. Fortunately this
clang-tidy rule provides check options to provide an allow-list of
variable names which better suits special cases such as these rather
than completely disabling the rule.

readability-isolate-declaration is allowed by the Google C++ Style Guide
which this project follows but is discouraged as part of the C++ Core
Guidelines. As the project as evolved over time, usage of this pattern
has diminished to a few instances and can therefore be refactored to
better align with the C++ Core Guidelines.