refactor: make gfx::Window API agnostic
Commit 8d369d9 refactored gfx::Window to use Vulkan types as part of its
API. Although this slightly simplifies the logic for interacting with
Vulkan APIs, it's considered questionable to depend on 3rd party types
when not required. For example, if this class wanted to be used in an
OpenGL renderer it would either force a dependency on Vulkan or require
refactoring. Therefore, gfx::Window is being refactored to avoid
dependencies on third party library types where avoidable by
conditionally compiling Vulkan dependent logic behind the
GLFW_INCLUDE_VULKAN compiler definition.