feat: add support for variable size index meshes
Previously, it was assumed that the index size for meshes would be 16
bit. While this is true for a majority meshes, this would fail for large
meshes with an index count greater than 65,536. This commit therefore
adds support for 32-bit index meshes which should be more than enough to
handle extremely large meshes.