feat: reinstate lighting using Phong shading
Commit f5c38a9 disabled the use of texture mapping and Phong shading
to focus on improving the overall design of the system. After a long
overdue departure involving adding Vulkan Memory Allocator, moving from
assimp to cgltf, and loading Basis Universal supercompressed KTX2
textures, it's time to reinstate the use of Phong shading. For now, the
only light source is hardcoded to be fixed to the origin of the camera.
Subsequent commits will address loading light data from a glTF file and
storing that in a uniform buffer.