fix: prevent dereferencing null scene data
The glTF specification does not require scene data to be defined. While
unlikely, it's possible to dereference potentially null scene data in
the current implementation. This commit addresses this by first checking
if the default scene is defined and if not use the first available scene
in the glTF scene list. In the event that no scene data is found, an
exception is thrown because the application cannot continue without a
valid scene graph.

This commit also features additional error validation improvements by
using assertions for scenarios which should never occur when processing
a valid glTF file. While not strictly necessary because it in part tests
the correctness of cgltf, these assertions are being added because (a)
they add not overhead for release builds and (b) document a requirement
from the glTF specification when invoking potentially unsafe behavior
(e.g., dereferencing a potential null pointer);