feat: add support for tangent vertex attributes
This commit enables support for tangent vertex attributes. To achieve
this, the logic for getting vertex attributes from glTF data needed to
be refactored to become more generic so that future vertex attributes
can be easily added and validated. As a consequence, error handling for
missing or invalid vertex attributes is much more robust.