refactor: rework device initialization
This commit is primarily aimed at improving and simplifying logical and
physical device initialization by moving queue family indices to the
physical device class and updating physical device selection based on
the existence of a valid set of queue family indices. This change also
removes use of a transfer queue family because it is currently unused
and can easily be added back if needed in the future.