refactor: add data view abstraction
This commit replaces uses of vk::ArrayProxy in public APIs in favor of
a dedicated data_view abstraction. This provides a few benefits over
vk::ArrayProxy such as providing an API to calculate its size in bytes
(similar to std::span) and avoid creating a dependency on Vulkan types
in public APIs to support additional graphics APIs in the future.