refactor: move main loop management to Engine::Run
The engine class currently exposes the Vulkan device for the sole
purpose of waiting for the device to become idle after the main loop.
A better approach is to instead provide a loop with a callback (inlined
as a lambda) to execute while it manages details related to the main
loop such as updating the delta time, polling for window events, etc.