refactor: improve vertex attribute validation
Previously, vertex attribute validation was done in the awkward way of
deconstructing vertex attributes into pairs to iterate over because
vertex attribute data types are not homogeneous. However, this can be
avoided by instead using a variadic template function that accepts an
arbitrary number of vertex attribute types to validate.