build: create symlink to assets directory
To ensure availability when running the game executable, assets are
copied from the source to binary directory using a custom cmake target.
While functional, copying assets can be time consuming if large which
directly impacts build times. Instead, a more lightweight approach is to
create a symlink to assets in the source directory.

It should be noted that creating symlinks does not work for distributing
a game, however, this project is still in the development stage and
therefore this approach is deemed sufficient.