feat: move camera transforms to uniform buffers
Camera transforms are currently stored as push constants in the form of
a model-view and projection transform. The use of a combined model-view
transform is problematic because the view transform is required for
certain algorithms such as normal mapping. Addition, this also will
simplify support of using multiple lights using the KHR_lights_punctual
glTF extension.

Therefore, this commit introduces the use of a uniform buffer to store
camera transforms. This commit also introduces an abstraction for
descriptor sets to manage descriptor set allocations and their lifetimes