refactor: improve material descriptor sets updates
Material descriptor sets are updated individually which less efficient
than updating all descriptor sets at once. Unfortunately, the current
design doesn't support doing this.

To achieve, material creation is being refactored to provide a reference
to a descriptor set so that it can easily reference the material images
it needs to be updated with.