refactor: move mesh implementation to a class
This commit provides a first pass at chipping away from the massive
amount of code in scene.cppm which is, in part, due to trying to
consolidate internal details related to glTF loading. However, the mesh
implementation is relatively straightforward and can be easily extracted
to a public class which only depends on Vulkan types (e.g, VkBuffer).