feat: add COLOR_0 vertex attribute support
This commit enables support for the COLOR_0 glTF vertex attribute which
acts as a linear multiplier with a material's base color when present.
To achieve this, vertex attribute validation was refactored to add
support for optional vertex attributes which are filled with a default
value when not specified in a glTF file.