fix: avoid potential memcpy out of bounds
When copying to a buffer, the size of the data to be copied in bytes
is verified to fit within the capacity of the buffer. However, there is
a bug where the amount of data copied to the buffer is the size of the
buffer and not the size of the data. This can result in copying beyond
the capacity of the original data pointer which can lead to undefined
behavior.