feat: add material properties uniform buffer
This commit enables support for copying material properties not derived
from a texture image in a uniform buffer for use in lighting equations
in the fragment shader. These initial properties come from the glTF 2.0
specification for PBR metallic-roughness materials that act as linear
multipliers with base color and metallic-roughness images.