fix: avoid race condition when creating samplers
When creating material samplers, an unordered map is used to keep track
of unique samplers based on sampler creation properties. Unfortunately
there is a bug where this unordered map is mutated from multiple threads
which is a race condition.

This bug can be partly attributed to the current design that allows for
a material texture image to be a variant of both a KTX texture and a
Vulkan device image. To fix this, the variant has been removed which
allows loading material KTX textures in isolation so that samplers
can then be created on a single thread when creating a material.