refactor: remove mesh class
The mesh class was prematurely added in an attempt to start breaking up
the glTF scene loading file. It unfortunately doesn't provide a great
abstraction because it leaks implementation details about its
corresponding material descriptor set layout. Therefore, this is being
merged back to be reworked as part of a more generic glTF loading
refactor.