feat: move lighting calculations to world space
Currently lighting calculations in the fragment shader are done in view
space to simplify getting the view direction. This also simplified
getting the light direction which was conveniently anchored to the
camera origin for testing purposes.

While there are some benefits to this approach, it unfortunately creates
a dependency on the camera which results in an additional transform to
view space for every object to be rendered in the scene which is
suboptimal when an object's position is static.

Therefore lighting calculations are being converted to world space. This
requires getting the camera position in the fragment shader to calculate
the world space view direction. Although this is encoded in the view
transform and could be easily extracted by the following equation:

  camera_pos = transpose(mat3(view_transform)) * -view_transform[3].xyz

this was deemed unnecessary to do in every vertex shader when this data
is readily available when recording draw commands. Therefore, this data
is being stored as a push constant with appropriate offsets for access
in the fragment shader.