refactor: add descriptor sets default constructor
This commit adds a default constructor to gfx::DescriptorSets to help
simplify usage for clients which may not have the data to immediately
allocate descriptor sets. Additionally, because it is a thin abstraction
over a list of descriptor sets, enabling default construction is a
natural API extension to representing an empty set. Lastly, class
invariants are still guaranteed through assertions or Vulkan validation.