refactor: rename scene to gltf scene
The original intention for gfx::Scene was to encapsulate all data needed
to render the scene and not depend on any one particular file format
loading implementation. Unfortunately, that is not the case and while
a larger refactor planned to split this up in the future, it is better
to rename the class now to represent its tight coupling with a glTF
loading implementation.