refactor: avoid runtime shader compilation
This project currently uses runtime shader compilation to convert GLSL
shaders to SPIR-V intermediate language. While this is useful for
dynamic shader generation, this project does not yet utilize such
functionality. It is therefore unnecessary to repeatedly compile GLSL
shaders each time the application is initialized.

To mitigate this problem, GLSL shader compilation is invoked by CMake as
a custom target that gets built as a dependency of the game executable
whenever a GLSL file is modified. Although currently unused, the GLSL
compiler implementation is retained for debugging purposes in addition
to potential use cases relating to dynamic shader compilation that may
arise in the near future.