refactor: simplify window APIs
gfx::Window has intentionally avoided using Vulkan and glm types as part
of its API to support compatibility with other graphics or math APIs in
the future. While there is merit to this approach, it is wildly out of
scope for this project and results in unnecessary intermediate types to
marshal data across API boundaries. Therefore, gfx::Window APIs have
been simplified to efficiently construct data in the format it will be
used.

This change also fixes a minor bug where the aspect ratio was being
computed with the outer window dimensions. Given that the swapchain uses
the framebuffer extent to establish its swapchain image dimensions and
the fact that we render to the entire framebuffer, it makes more sense
to compute the aspect ratio using the framebuffer dimensions as well.