refactor: restrict window creation to fullscreen
The current window implementation attempts to simulate fullscreen by
creating a window with its maximum dimensions (up to 4k resolution which
is admittedly a hack). However, there are several benefits to creating a
window in fullscreen mode including better performance and a simplified
implementation due to restrictions on window resizing.

Although enabling windowed mode support is desirable, it requires a more
thought out approach which can gracefully handle window resizing; a
currently unsupported scenario. Therefore, support for creating windows
in anything other than fullscreen is being dropped with the expectation
of revising in the future when such functionality is needed.