refactor: improve read file error handling
The current shader module implementation contains a minor bug with
reading SPIR-V and GLSL files where it fails to check the state of the
file stream after reading the file. Although unlikely to occur after
verifying the file opened successfully, it can still become corrupted
while reading due to external modifications such as being deleted. It is
therefore considered a best practice to check the state of an `ifstream`
after reading its contents.

However this approach does not scale well because it requires explicitly
checking the state of an `ifstream` after any API call which can quickly
lead to a rats nest of exception handling. Instead, a better approach is
to enable exceptions for an `ifstream` so that it will automatically
throw anytime its `failbit` or `badbit` flags are set. To ensure error
messages are still actionable, `std::throw_with_nested` is used to chain
the original `std::ios::failure` exception in a `std::runtime_error` to
indicate what file could not be read. Lastly, to correctly report nested
error messages, `main.cpp` was refactored to support recursive exception
handling.