refactor: improve pass by value/reference usages
In general, pass by value is preferred for small data types that are
cheap to copy. On x64 architecture, this includes anything less than the
size of a pointer which is one word size or 8 bytes.

This commit implements a review of this policy to ensure data types less
than 8 bytes are passed by value whereas types larger than 8 bytes are
passed by reference.