refactor: use ISO spherical coordinate convention
There are several different conventions for how to represent spherical
coordinates. Although they all use the triple (r, theta, phi), different
conventions assign different angles to theta and phi. This project uses
theta for the azimuthal angle and phi for the polar angle. However, an
alternative convention used in physics which has been established as an
ISO standard is instead assign phi to the azimuthal angle and theta to
the polar angle. While there are merits to the current approach (notably
maintaining consistency with 2D polar angles), it is better to maintain
consistency with an established ISO standard. Therefore, this commit
swaps the meaning of each angle to algin with this standard.