refactor: improve camera translation performance
When processing keyboard input, the camera is translated for each WASD
key that is pressed which is inefficient because the view transform is
calculated as subroutine of camera translation. Instead, a better
approach is to incrementally construct a single translation vector over
all keyboard input so the camera can be translated once.