refactor: convert game start to a free function
The current API for game initialization is awkward because it requires
constructing a game object followed by an immediate need to invoke its
start API to run the core engine loop. While there are merits to this
approach, its flexibility is  unnecessary and is not foreseen to be
needed in the near future. Therefore, the game object is being removed
in favor of a single free function API that handles both game
initialization and starting the main engine loop.