refactor: derive euler angles from view transform
Internally, the camera implementation uses spherical coordinates to
represent its orientation. While efficient for applying rotations, it
is problematic because it requires constructing the view transformation
matrix on demand which becomes inefficient when reconstructed multiple
times in the main render loop where performance is critical.

One approach to mitigate this is to construct the view transform matrix
anytime the camera is translated or rotated so it can later be accessed
in constant time. However, this is also suboptimal because it requires
caching the view transformation matrix which creates redundancy with the
camera position and orientation that risks becoming out of sync.

Because the view transformation matrix is what is ultimately needed by
in the vertex shader, a better approach is to instead derive the camera
position and orientation from the view transformation matrix when
applying camera transformations. For view-space translations, this is
trivial since this vector is already represented in the fourth column of
the view transform matrix. For rotations, the pitch and yaw euler angles
can be derived from the upper 3x3 submatrix of the view transform that
represents its orientation. Following a similar approach as spherical
coordinates, these angles can be cumulatively updated to create a Euler
rotation matrix that represents the new camera orientation. One gotcha
to be aware of is that the view transformation translation vector also
encodes the previous rotation so it must also be inverted to get the
camera's world position so that the new euler rotation can be applied.

Lastly, this refactor also enabled retrieving the projection transform
in constant time which previously followed a similar approach of the
view transformation where it was computed on demand from the camera's
view transform. This was particularly inefficient since there was no way
for the projection transform to change as a result of the view frustum
changing. This further enables different projection methods (e.g.,
orthographic, infinite perspective) without conditional branching in the
main render loop which would have been necessary with the previous
approach.